make ShowEntry a separate component, render on '/show' route
