+++ src/components/auth/calendar-view.jsx
@@ -5,6 +5,7 @@ import { Link, Switch, Route } from 'react-router-dom';
 import Calendar from './calendar';
 import ShowEntries from './show-entries';
 import Search from './search';
+import ShowEntry from './show-entry';
 import { setNowDate } from '../../actions';
 
 class CalendarView extends Component {
@@ -45,6 +46,7 @@ class CalendarView extends Component {
                 <Switch>
                   <Route exact path="/" component={ShowEntries} />
                   <Route path="/search" component={Search} />
+                  <Route path="/show" component={ShowEntry} />
                 </Switch>
             </Segment>
           </Grid.Column>
